requirements.txt
python-dotenv
groq
tweepy
pillow
requests
python-crontab
selenium